backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.0
pymysql==1.1.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
openai==1.0.0 